organized the react files
